requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
djangorestframework==3.15.2